AWS to Cloudinary

+++ src/pages/api/post/[postId].ts
@@ -1,7 +1,13 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import nc from 'next-connect';
 import { prisma } from "../../../../lib/prisma";
-import AWS from 'aws-sdk';
+import nc from 'next-connect';
+import cloudinary from 'cloudinary';
+
+cloudinary.v2.config({
+  cloud_name: process.env.CLOUDINARY_NAME,
+  api_key: process.env.CLOUDINARY_KEY,
+  api_secret: process.env.CLOUDINARY_SECRET,
+});
 
 const apiRoute = nc<NextApiRequest, NextApiResponse>({
 
@@ -17,11 +23,6 @@ const apiRoute = nc<NextApiRequest, NextApiResponse>({
   
 })
 .delete(async( req, res ) => {
-  const s3 = new AWS.S3({
-    accessKeyId: process.env.AWS_ID,
-    secretAccessKey: process.env.AWS_SECRET,
-    signatureVersion: 'v4'
-})
 
   const { postId } = req.query;
   const id = parseInt(postId);
@@ -34,22 +35,18 @@ const apiRoute = nc<NextApiRequest, NextApiResponse>({
   })
 
   if(file){
-    const params = {
-      Bucket: process.env.AWS_BUCKET,
-      Key: file.video
+    const del = await cloudinary.v2.uploader.destroy(file.publicId,  {type : 'upload', resource_type : 'video'})
+
+    var result = null;
+    if(del.result === 'ok'){
+      result = await prisma.post.delete({
+        where: {
+          id: id
+        }
+      })
     }
 
-    s3.deleteObject(params, (err, data) => {
-      if(err) res.status(500).send({Error: "couldn't delete the post"});
-    })
-
-    const result = await prisma.post.delete({
-      where: {
-        id: id
-      }
-    })
-
-    if(result) res.status(200).send(result)
+    if(result && del.result === 'ok') res.status(200).send(result)
     else res.status(500).send({Error: "couldn't delete the post"})
   }
   else res.status(500).send({Error: "couldn't delete the post"})
